services/data-ingestion/requirements.txt
requests==2.31.0
kafka-python-ng==2.2.3
flask==2.3.3
pytest==7.4.2
pytest-mock==3.11.1
responses==0.23.3